update del_meta_data function
The key replacement "deleted_by_bidsonym" tool place in every json file as the function was set up in the wrong way. Now there's a check that overwrites only if the key exists (in the previous version a key which value should be deleted was added in cases where it didn't exist).