Return lists of defaced t1w and t2w and define it for t2w
Otherwise if no t1w files found, the variable t1w would not get populated in
the loop at all and cause the crash.  But it also does not make much sense to
return only the last hit in the loop. Checking the workflow, it seems
outputs are not used at all anyways, so we are gonna try to return a list